requirements.txt
asgiref==3.4.1; python_version >= '3.6'
beautifulsoup4==4.10.0; python_version >= '3.1'
bs4==0.0.1
django-cors-headers==3.10.0
django-phonenumber-field[phonenumbers]==6.0.0
django-service-objects==0.7.0
django==3.2.9
djangorestframework==3.12.4
phonenumbers==8.12.38
pillow==8.4.0
pytz==2021.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.1; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
